+++ examples/carousel.html
@@ -72,7 +72,7 @@
   </style>
 
   <div class="parent" id="parent">
-    <h2>Carousel A - Using PhysicalSlide transition</h2>
+    <h2>Carousel A - Using DraggableSlide transition</h2>
     <div class="carousel">
       <div class="slide">1</div>
       <div class="slide">2</div>
