+++ examples/carousel.js
@@ -3,7 +3,7 @@
  * updated by the scroll position of a position sticky based container.
  */
 import {Carousel} from '../lib/ui/carousel/carousel';
-import {PhysicalSlide} from '../lib/ui/carousel/physical-slide/physical-slide';
+import {DraggableSlide} from '../lib/ui/carousel/draggable-slide/draggable-slide';
 import {CarouselSynchronizer} from '../lib/ui/carousel/carousel-synchronizer';
 
 export default class CarouselSample {
@@ -14,7 +14,7 @@ export default class CarouselSample {
         Array.from(document.querySelectorAll('.carousel .slide')),
         {
           allowLooping: true,
-          transition: new PhysicalSlide(),
+          transition: new DraggableSlide(),
         }
     );
     const b = new Carousel(
