+++ src/ui/carousel/draggable-slide/matrix-service.ts
